separate example data
